+++ ogc/features/datasources/geopackage.go
@@ -17,7 +17,7 @@ func (GeoPackage) Close() {
 	// TODO: clean up DB connection to gpkg
 }
 
-func (GeoPackage) GetFeatures(collection string, cursor int, limit int) (*domain.FeatureCollection, domain.Cursor) {
+func (GeoPackage) GetFeatures(collection string, cursor int64, limit int) (*domain.FeatureCollection, domain.Cursor) {
 	// TODO: not implemented yet
 	log.Printf("TODO: return data from gpkg for collection %s using cursor %d with limt %d",
 		collection, cursor, limit)
